hive/cell/health: don't warn when reporting on stopped reporter.
Currently in cases where reporters emit status on stopped reporters, a warning is logged.
This causes failures in CI related to endpoints, as those close their reporters after endpoint is deleted.

This scenario is probably not a valuable log for users and generally such occurances should be harmless.
This moves the log to debug level.

Fixes: #31147

Signed-off-by: Tom Hadlaw <tom.hadlaw@isovalent.com>